temporal/pyproject.toml
[project]
name = "code-doc-assistant-temporal"
version = "1.0.0"
description = "Code Documentation Assistant Temporal Worker"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Code Doc Assistant Team"}
]

dependencies = [
    # Temporal SDK
    "temporalio>=1.21.1",
    # Code Parsing
    "tree-sitter>=0.25.2",
    "tree-sitter-languages>=1.10.2",
    # Vector Store
    "chromadb>=1.4.1",
    # Embeddings
    #"jinahub[jinaai]>=0.1.0",
    # OpenAI (fallback)
    "openai>=2.15.0",
    # HTTP Client
    "httpx>=0.28.1",
    # Utilities
    "gitpython>=3.1.46",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.11",
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

#[tool.hatch.build.targets.wheel]
#packages = ["workflows", "activities"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
